+++ back/src/main/java/com/stock/survive/security/refresh/RefreshTokenStore.java
@@ -3,7 +3,7 @@ package com.stock.survive.security.refresh;
 import java.time.Duration;
 
 public interface RefreshTokenStore {
-    String issue(Integer userNo, Duration ttl);
-    Integer verify(String raw);
+    String issue(Long userNo, Duration ttl);
+    Long verify(String raw);
     void revoke(String raw);
 }
